misc/HtmlTestSuite/results/test-0764.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="https://drafts.css-houdini.org/css-properties-values-api-1" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../resources/utils.js"/>
</head><body><div id="target"/>
<script>

test(() =&gt; {
  CSS.registerProperty({
    name: "--my-length",
    syntax: "&lt;length&gt;",
    inherits: false,
    initialValue: "0px"
  });

  target.style.marginLeft = "var(--my-length)";

  const duration = 1000;
  const animation = target.animate({ "--my-length": "100px" }, duration);
  animation.pause();
  animation.currentTime = duration / 4;

  assert_equals(getComputedStyle(target).marginLeft, "25px");
}, "Animating a non-inherited CSS variable is reflected on a standard property using that variable as a value");

</script></body></html>
